pyproject.toml
[tool.poetry]
name = "rpa-framework"
version = "0.10.1"
description = "A collection of tools and libraries for RPA"
authors = [
	"RPA Framework <rpafw@robocorp.com>",
]
license = "Apache-2.0"
readme = "README.rst"

homepage = "https://rpaframework.org/"
documentation = "https://rpaframework.org/"
repository = "https://github.com/robocorp/rpa-framework"

keywords = ["robotframework", "rpa", "automation"]

classifiers = [
	"License :: OSI Approved :: Apache Software License",
	"Development Status :: 3 - Alpha",
	"Operating System :: OS Independent",
	"Intended Audience :: Developers",
	"Topic :: Software Development :: Libraries :: Python Modules",
	"Topic :: Software Development :: Libraries",
	"Framework :: Robot Framework :: Library",
	"Framework :: Robot Framework",
	"Programming Language :: Python :: 3.6",
	"Programming Language :: Python :: 3.7",
	"Programming Language :: Python :: 3.8",
]

packages = [
	{ include = "RPA", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.6"
dataclasses = {version = "^0.7", python=">=3.6,<3.7"}
robotframework-seleniumlibrary = "^4.1.0"
robotframework-sapguilibrary = {version = "^1.1", platform = "win32"}
robotframework-requests = "^0.6.5"
webdrivermanager = ">=0.7.4"
pywinauto = {version = "^0.6.8", platform = "win32", python = "!=3.8.1,!=3.7.6"}
pywin32 = {version = "^227", platform = "win32", python = "!=3.8.1,!=3.7.6"}
psutil = {version = "^5.7.0", platform = "win32"}
clipboard = "^0.0.4"
exchangelib = "^3.1.1"
fpdf = "^1.7.2"
xlrd = "^1.2.0"
xlwt = "^1.3.0"
xlutils = "^2.0.0"
openpyxl = "^3.0.3"
pillow = "^7.0.0"
pyscreenshot = "^1.0"
opencv-python = {version = "^4.2.0", optional = true}
numpy = {version = "^1.18.2", optional = true}
robotframework-databaselibrary = "^1.2.4"
simple_salesforce = "^1.0.0"
netsuitesdk = "^1.1.0"
"pdfminer.six" = "^20200402"
pypdf2 = "^1.26.0"
tweepy = "^3.8.0"
robotframework = "^3.2.1"
boto3 = {version = "^1.13.4", optional = true}
google-api-python-client = "^1.8.3"


[tool.poetry.dev-dependencies]
black = { version = "*", allow-prereleases = true }
flake8 = "^3.7.9"
pylint = "^2.4.4"
pytest = "^5.2"
sphinx = "^2.3.1"
sphinx_rtd_theme = "^0.4.3"
robotframeworklexer = "^1.1"
sphinx-jekyll-builder = "^0.3.0"
mock = "^4.0.2"

[tool.poetry.extras]
cv = ["opencv-python", "numpy"]
aws = ["boto3"]

[tool.black]
target-version = ["py36", "py37", "py38"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
